+++ CHANGES.md
@@ -35,6 +35,7 @@ Changes in 10.0
 - HTML add label to select in ActivityReport.php
 - Use Currency() function instead of number_format() in TransactionsReport.php
 - HTML remove line-break in Warning/Minimum columns in Reminders.php
+- HTML CSS make Daily Menus calendar coherent with School Calendar in DailyMenus.php
 - Fix dummy day (year month date) set to 28 for February in Dashboard.inc.php
 
 Changes in 9.3